# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/_internal/v1/trigger_codegen": {
      "post": {
        "operationId": "trigger-codegen",
        "responses": {
          "200": {
            "description": "Codegen triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerCodegenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "TriggerCodegenResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}